refactor: extract video optimization queue service
